requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3

# Authentication
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1

# Database
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pdfkit==1.0.0
pdfrw==0.4.0
Pillow>=9.5.0
pytesseract==0.3.10
pdf2docx==0.5.6
fitz==0.0.1.dev2
img2pdf==0.4.4
pdfminer.six==20221105
fpdf==1.7.2
pytesseract==0.3.10

# Document Processing
python-docx==0.8.11
docx2pdf==0.1.8
markdown==3.4.1
html2text==2020.1.16
beautifulsoup4==4.11.2
PyMuPDF==1.25.4
reportlab==4.3.1
# Temporarily comment out lxml until Python 3.13 wheels are available
# lxml==4.9.2

# Audio Processing
soundfile==0.12.1
pydub==0.25.1; python_version < '3.13'
# Remove simpleaudio as it's not compatible
numpy>=1.24.0  # Required for soundfile
ffmpeg-python==0.2.0

# Utilities
python-magic-bin==0.4.14; platform_system == 'Windows'
python-magic==0.4.27; platform_system != 'Windows'
requests==2.28.2
tqdm==4.65.0
uuid==1.30
python-dotenv==1.0.0
gunicorn==20.1.0
Babel==2.11.0

# Testing
pytest==7.2.2
coverage==7.2.1